refactor(study): change project package name
상단 패키지의 이름을 변경하였습니다
