+++ src/main/java/com/mogaco/project/web/WebController.java
@@ -1,4 +1,4 @@
-package com.mogaco.study.web;
+package com.mogaco.project.web;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
