+++ src/test/java/com/mogaco/project/StudyApplicationTests.java
@@ -1,4 +1,4 @@
-package com.mogaco.study;
+package com.mogaco.project;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
